# source/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WNR-Audit API — Microsoft tenant security auditing
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: Listar servidores on-premise cadastrados
      responses:
        "200":
          description: Lista de servidores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServerDevice"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createServer
      tags: [servers]
      summary: Cadastrar novo servidor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerBody"
      responses:
        "201":
          description: Servidor criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDevice"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /servers/{deviceId}:
    parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getServer
      tags: [servers]
      summary: Detalhes do servidor
      responses:
        "200":
          description: Servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerDeviceDetails"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Remover servidor
      responses:
        "204":
          description: Removido
        "404":
          $ref: "#/components/responses/NotFound"
  /servers/{deviceId}/scans:
    parameters:
      - name: deviceId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: listServerScans
      tags: [servers]
      summary: Listar varreduras do servidor
      responses:
        "200":
          description: Varreduras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceScan"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: startServerScan
      tags: [servers]
      summary: Iniciar varredura no servidor
      responses:
        "201":
          description: Varredura iniciada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceScan"
        "404":
          $ref: "#/components/responses/NotFound"
  /server-scans/{scanId}:
    parameters:
      - name: scanId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getServerScan
      tags: [servers]
      summary: Detalhes de varredura de servidor
      responses:
        "200":
          description: Detalhes da varredura
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceScanDetails"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Listar todos os usuários (apenas super admin)
      responses:
        "200":
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/invite:
    post:
      operationId: inviteUser
      tags: [users]
      summary: Convidar novo usuário por e-mail (cria registro pendente)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserBody"
      responses:
        "201":
          description: Usuário convidado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /users/{userId}/approve:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: approveUser
      tags: [users]
      summary: Aprovar acesso de um usuário pendente
      responses:
        "200":
          description: Usuário aprovado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}/block:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: blockUser
      tags: [users]
      summary: Bloquear acesso de um usuário
      responses:
        "200":
          description: Usuário bloqueado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Excluir um usuário e todos os dados associados
      responses:
        "204":
          description: Usuário excluído
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict (already exists)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden — caller does not have required role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Severity:
      type: string
      enum: [critical, high, medium, low, info]
    FindingStatus:
      type: string
      enum: [open, ignored, resolved]
...
    ScanStatus:
      type: string
      enum: [pending, running, completed, failed]
...
    UserStatus:
      type: string
      enum: [pending, active, blocked]
    UserRole:
      type: string
      enum: [super_admin, user]
...
    ServerOs:
      type: string
      enum:
        - windows-server
        - ubuntu
        - debian
        - rhel
        - centos
        - rocky-linux
        - suse
        - other
    DeviceType:
      type: string
      enum:
        - firewall
        - server
        - external
...
    ServerDevice:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        name:
          type: string
        hostname:
          type: string
        os:
          $ref: "#/components/schemas/ServerOs"
        osVersion:
          type: string
          nullable: true
        ipAddress:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        lastScanAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, customerId, name, hostname, os, createdAt, updatedAt]
    ServerDeviceDetails:
      allOf:
        - $ref: "#/components/schemas/ServerDevice"
        - type: object
          properties:
            recentScans:
              type: array
              items:
                $ref: "#/components/schemas/DeviceScan"
          required: [recentScans]
    CreateServerBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        hostname:
          type: string
          minLength: 1
        os:
          $ref: "#/components/schemas/ServerOs"
        osVersion:
          type: string
        ipAddress:
          type: string
        role:
          type: string
        notes:
          type: string
      required: [name, hostname, os]
    DeviceScan:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        deviceType:
          $ref: "#/components/schemas/DeviceType"
        status:
          $ref: "#/components/schemas/ScanStatus"
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        totalChecks:
          type: integer
        passedChecks:
          type: integer
        failedChecks:
          type: integer
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
      required:
        - id
        - deviceId
        - deviceType
        - status
        - startedAt
        - totalChecks
        - passedChecks
        - failedChecks
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
    DeviceScanDetails:
      allOf:
        - $ref: "#/components/schemas/DeviceScan"
        - type: object
          properties:
            deviceName:
              type: string
            findings:
              type: array
              items:
                $ref: "#/components/schemas/DeviceFinding"
            controls:
              type: array
              items:
                $ref: "#/components/schemas/DeviceControl"
          required: [deviceName, findings, controls]
    DeviceFinding:
      type: object
      properties:
        id:
          type: string
        deviceScanId:
          type: string
        controlId:
          type: string
        title:
          type: string
        category:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        status:
          $ref: "#/components/schemas/FindingStatus"
        affectedResource:
          type: string
          nullable: true
        detectedAt:
          type: string
          format: date-time
      required:
        - id
        - deviceScanId
        - controlId
        - title
        - category
        - severity
        - status
        - detectedAt
    DeviceControl:
      type: object
      properties:
        controlId:
          type: string
        title:
          type: string
        category:
          type: string
        severity:
          $ref: "#/components/schemas/Severity"
        affectedResource:
          type: string
          nullable: true
        recommendation:
          type: string
        status:
          type: string
          enum: [passed, failed]
        findingId:
          type: string
          nullable: true
      required: [controlId, title, category, severity, recommendation, status]
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
          nullable: true
        organizationName:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        role:
          $ref: "#/components/schemas/UserRole"
        approvedAt:
          type: string
          format: date-time
          nullable: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, status, role, createdAt]
    InviteUserBody:
      type: object
      properties:
        email:
          type: string
          format: email
        fullName:
          type: string
      required: [email]
...