# source/lib/api-zod/src/generated/types/createServerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WNR-Audit API — Microsoft tenant security auditing
 * OpenAPI spec version: 0.1.0
 */
import type { ServerOs } from "./serverOs";

export interface CreateServerBody {
  /** @minLength 2 */
  name: string;
  /** @minLength 1 */
  hostname: string;
  os: ServerOs;
  osVersion?: string;
  ipAddress?: string;
  role?: string;
  notes?: string;
}