# source/lib/api-zod/src/generated/types/serverDevice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WNR-Audit API — Microsoft tenant security auditing
 * OpenAPI spec version: 0.1.0
 */
import type { ServerOs } from "./serverOs";

export interface ServerDevice {
  id: string;
  customerId: string;
  name: string;
  hostname: string;
  os: ServerOs;
  osVersion?: string | null;
  ipAddress?: string | null;
  role?: string | null;
  notes?: string | null;
  lastScanAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}