# source/lib/api-zod/src/generated/types/serverDeviceDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WNR-Audit API — Microsoft tenant security auditing
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceScan } from "./deviceScan";
import type { ServerDevice } from "./serverDevice";

export type ServerDeviceDetails = ServerDevice & {
  recentScans: DeviceScan[];
};