# source/lib/api-zod/src/generated/types/serverOs.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WNR-Audit API — Microsoft tenant security auditing
 * OpenAPI spec version: 0.1.0
 */

export type ServerOs = (typeof ServerOs)[keyof typeof ServerOs];

export const ServerOs = {
  "windows-server": "windows-server",
  ubuntu: "ubuntu",
  debian: "debian",
  rhel: "rhel",
  centos: "centos",
  "rocky-linux": "rocky-linux",
  suse: "suse",
  other: "other",
} as const;